Fix #7189.
Former-commit-id: c0a2a5dd682668e2364c05e7d6bdbe022e5fa66b

+++ source/ch/cyberduck/core/io/BandwidthThrottle.java
@@ -101,7 +101,7 @@ public final class BandwidthThrottle {
     /**
      * No throttling
      */
-    public static final float UNLIMITED = -1;
+    public static final int UNLIMITED = -1;
 
     /**
      * Bytes per second allowed
